% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standAge.R
\name{ABovE_CaNFIR_standAge}
\alias{ABovE_CaNFIR_standAge}
\title{Estimate cell-specific stand age}
\usage{
ABovE_CaNFIR_standAge(
  agbfiles,
  distfiles,
  age_mosaic,
  tilenames,
  years,
  cl = NULL
)
}
\arguments{
\item{agbfiles, distfiles}{character vector of filenames specifying ABoVE AGB and disturbance
raster tiles.}

\item{age_mosaic}{character string specifying the filepath to CaNFIR stand age mosaic.}

\item{tilenames}{character vector of output tile directories.}

\item{years}{integer vector of data years.}

\item{cl}{cluster object. if \code{NULL} (default), a cluster will be created using up to
\code{length(intervals)} number of CPU cores for parallel computation.
Users can pass their own \code{cl} object or specify option
\code{parallelly.availableCores.fallback} to reduce the number of cores used.
See \code{?parallelly.options}.}
}
\value{
Invisibly, a list of output files created by the function.
Invoked for the side effect of writing these raster tiles to disk.
}
\description{
Uses a combination of ABoVE Disturbance Agents and the CaNFIR stand age mosaic (kNN 2020)
}
