% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBtrends-package.R
\docType{package}
\name{AGBtrends-package}
\alias{AGBtrends}
\alias{AGBtrends-package}
\title{AGBtrends: Utilities for Analyzing Trends in Above Ground Biomass}
\description{
Tools for analyzing above-ground biomass (AGB) and disturbance timeseries datasets created through the Arctic-Boreal Vulnerability Experiment ('ABoVE') research project (\url{https://daac.ornl.gov/cgi-bin/dataset_lister.pl?p=34}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PredictiveEcology/AGBtrends}
  \item Report bugs at \url{https://github.com/PredictiveEcology/AGBtrends/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca}

Authors:
\itemize{
  \item Tyler D Rudolph \email{tyler.rudolph@nrcan-rncan.gc.ca}
  \item Céline Boisvenue \email{celine.boisvenue@nrcan-rncan.gc.ca}
}

}
\keyword{internal}
