% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMosaics.R
\name{buildMosaics}
\alias{buildMosaics}
\title{Build raster mosaics for each time interval}
\usage{
buildMosaics(type, intervals, paths, cl = NULL)
}
\arguments{
\item{type}{character string specifying the mosaic type to build.
one of \verb{"age", "landCover", "slope", "sample_size"}.}

\item{intervals}{named list of time intervals over which to build mosaics}

\item{paths}{named list of directory paths, specifying paths for at least:
\code{scratch} and \code{tiles}.}

\item{cl}{cluster object.}
}
\value{
character vector of output raster file names (\code{.tif}).
Invoked for side effects of building and writing raster mosaics to disk.
}
\description{
Build raster mosaics for each time interval
}
