% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getNumCores}
\alias{getNumCores}
\title{Determine the number of available cores for parallel processing}
\usage{
getNumCores(maxCores = 0.5)
}
\arguments{
\item{maxCores}{the maximum number or proportion of cores to use.}
}
\value{
integer
}
\description{
Will return at least one core, up to a maximum of \code{prop} the available CPU cores,
or option \code{parallelly.availableCores.fallback}, if specified.
}
