% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.R
\name{gwr}
\alias{gwr}
\alias{gwrt}
\title{Geographically Weighted Regression}
\usage{
gwr(tileDirs, type = NULL, cores = 1L, prefix = NULL, lyrs = NULL)

gwrt(tileDirs, type = NULL, cores = 1L, intervals = NULL)
}
\arguments{
\item{tileDirs}{character. vector of directory paths to input raster tiles.}

\item{type}{character. one of \code{"nsamp"} or \code{"slope"}.}

\item{cores}{positive integer. number of cores to use for parallel processing.}

\item{prefix}{character. output filename prefix.
default \code{NULL} automatically sets this based on \code{type}.}

\item{lyrs}{positive integer. indices corresponding to raster layers for specific \code{intervals}.}

\item{intervals}{named list. timeseries intervals, e.g. \code{list(t1 = 1:5, t2 = 6:10, ...)}.}
}
\value{
character. vector of file paths corresponding to output rasters.
}
\description{
Estimate cell-wise linear regression coefficients for undisrupted time series
(a.k.a. "local" or "geographically weighted regression (GWR)").
}
\details{
\code{gwr()} operates on the entire timeseries, whereas \code{gwrt()} operates on the \code{intervals}.
}
