% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-zoneStats.R
\name{plotZoneStats}
\alias{plotZoneStats}
\title{Plot summary statistics}
\usage{
plotZoneStats(file2plot = NULL, weighted = TRUE, out = 1)
}
\arguments{
\item{file2plot}{TODO}

\item{weighted}{TODO}

\item{out}{TODO}
}
\value{
TODO
}
\description{
Plot summary statistics
}
