% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-zoneStats.R
\name{plotZoneStatsIntervals}
\alias{plotZoneStatsIntervals}
\title{Plot zone stats intervals}
\usage{
plotZoneStatsIntervals(
  files2plot = NULL,
  weighted = TRUE,
  xVar = "ageClass",
  groupVar = "tp",
  catVar = NULL,
  ptype = 1,
  plotResult = TRUE
)
}
\arguments{
\item{files2plot}{TODO}

\item{weighted}{TODO}

\item{xVar}{TODO}

\item{groupVar}{TODO}

\item{catVar}{TODO}

\item{ptype}{integer indicating plot type: 1 is XXX; 2 i- YYY. TODO}

\item{plotResult}{TODO}
}
\value{
TODO
}
\description{
TODO DESCRIPTION NEEDED
}
