% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepZones.R
\name{prepZones}
\alias{prepZones}
\title{Prepare categorical zones for comparative analysis}
\usage{
prepZones(
  zoi,
  field = "ECOZONE",
  ageClass,
  fileID = "WBI_ecozone_t1",
  cropRaster = NULL,
  destinationPath,
  overwrite = TRUE
)
}
\arguments{
\item{zoi}{zone-of-interest \code{sf} object created using \code{createAnalysisZones()},
or a character giving a file path to such an object that can be
read using \code{st_read()}.}

\item{field}{character string giving the field (column) in \code{zoi} to use for
comparative analyses.}

\item{ageClass}{\code{SpatRaster} object representing age classes}

\item{fileID}{TODO: e.g. one of \code{"WBI_ecozones"}}

\item{cropRaster}{\code{SpatRaster} to use to crop \code{ageClass}}

\item{destinationPath}{character string providing the path to a directory where
downloaded data or prepared object(s) will be saved.}

\item{overwrite}{logical. should the file be overwritten if it exists?}
}
\value{
(invisibly) character string giving the file path to the output raster file.
}
\description{
Create unique categorical zones for desired time period
(e.g. study area ZOI x age class).
}
