% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{data}
\name{Canada_Albers_Equal_Area_Conic}
\alias{Canada_Albers_Equal_Area_Conic}
\title{\code{Canada_Albers_Equal_Area_Conic} projection}
\format{
An object of class \code{character} of length 1.
}
\usage{
Canada_Albers_Equal_Area_Conic
}
\description{
Unknown EPSG code associated with this projection, which is used for the ABoVE datasets.
}
\keyword{datasets}
