% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.R
\name{ts_slope}
\alias{ts_slope}
\alias{ts_nsamp}
\title{Calculate slopes and number of samples of timeseries}
\usage{
ts_slope(x)

ts_nsamp(x)
}
\arguments{
\item{x}{vector of timeseries values}
}
\value{
TODO
}
\description{
\itemize{
\item \code{ts_slope()} derives slope of numerical vector across a time series;
\item \code{ts_nsamp()} calculates the number of non-\code{NA} values in a sample;
}
}
