% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-zoneStats.R
\name{tref}
\alias{tref}
\title{Timeseries interval reference}
\usage{
tref(timeint, years)
}
\arguments{
\item{timeint}{timeseries intervals as a named list with years corresponding to each interval
(named \code{t1}, \code{t2}, \code{t3}, etc.).}

\item{years}{integer vector of timeseries years}
}
\value{
data.frame with columns \code{tp} (time period), \code{yr1} (first years of each time period),
and \code{t2} (the last years of each time period)
}
\description{
Timeseries interval reference
}
\examples{
timeint <- list(
  t1 = 1:5, t2 = 6:10, t3 = 11:15, t4 = 16:20, t5 = 21:25, t6 = 26:31
)
years <- unname(unlist(timeint)) + 1983L
tref(timeint, years)
}
