% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ts_slope}
\alias{ts_slope}
\title{Derive slope of numerical vector across a time series}
\usage{
ts_slope(x)
}
\arguments{
\item{x}{vector of timeseries values}
}
\value{
TODO
}
\description{
Derive slope of numerical vector across a time series
}
