% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoneStats.R
\name{zoneStats}
\alias{zoneStats}
\title{Calculate summary statistics by categorical zone of interest}
\usage{
zoneStats(
  slopeRaster = NULL,
  weightRaster = NULL,
  zoneRaster = NULL,
  cropRaster = NULL,
  maskRaster = NULL,
  fileID = NULL,
  destinationPath = NULL
)
}
\arguments{
\item{slopeRaster}{TODO}

\item{weightRaster}{TODO}

\item{zoneRaster}{TODO}

\item{cropRaster}{TODO}

\item{maskRaster}{TODO}

\item{fileID}{TODO: e.g. one of \code{"WBI_ecozones"}}

\item{destinationPath}{character string providing the path to a directory where
downloaded data or prepared object(s) will be saved.}
}
\value{
(invisibly) character string corresponding to the file path of the saved raster.
}
\description{
Weighted mean & standard deviation of slopes by age class, time period,
and study area zones of interest.
}
