% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMosaics.R
\name{buildMosaics}
\alias{buildMosaics}
\title{Build raster mosaics for each time interval}
\usage{
buildMosaics(intervals, paths, cores)
}
\arguments{
\item{intervals}{named list of time intervals over which to build mosaics}

\item{paths}{named list of directory paths, specifying paths for at least:
\code{scratch} and \code{tiles}.}

\item{cores}{integer specifying the number of cores to use for parallel computations}
}
\value{
\code{NULL}, invisibly. Invoked for side effects of building and writing raster mosaics
to disk.
}
\description{
Build raster mosaics for each time interval
}
