% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMosaics.R
\name{buildMosaics}
\alias{buildMosaics}
\title{Build raster mosaics for each time interval}
\usage{
buildMosaics(type, intervals, src, dst, cl = NULL)
}
\arguments{
\item{type}{character string specifying the mosaic type to build.
one of \verb{"age", "LandCover", "LandCover_Simplified, "slope", "sample_size"}.}

\item{intervals}{named list of time intervals over which to build mosaics}

\item{src, dst}{character. vector of directory paths to source and destination raster files.}

\item{cl}{cluster object. if \code{NULL} (default), a cluster will be created using up to
\code{length(intervals)} number of CPU cores for parallel computation.
Users can pass their own \code{cl} object or specify option
\code{parallelly.availableCores.fallback} to reduce the number of cores used.
See \code{?parallelly.options}.}
}
\value{
character vector of output raster file names (\code{.tif}).
Invoked for side effects of building and writing raster mosaics to disk.
}
\description{
Build raster mosaics for each time interval
}
