% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{data}
\name{Canada_Albers_Equal_Area_Conic}
\alias{Canada_Albers_Equal_Area_Conic}
\title{\code{Canada_Albers_Equal_Area_Conic} projection}
\format{
An object of class \code{character} of length 1.
}
\usage{
Canada_Albers_Equal_Area_Conic
}
\description{
Unknown EPSG code associated with this projection, which is used for the ABoVE datasets.
}
\details{
This is the character representation of the CRS, which may need to be converted for
use with geospatial operations (e.g., using \code{sf::st_crs()} or \code{terra::crs()}).
}
\keyword{datasets}
